int32_t=>int64_t

+++ oneflow/core/register/blob.cpp
@@ -23,9 +23,9 @@ void Blob::Init(Regst* regst, const RtBlobDesc* blob_desc, char* header_ptr, cha
   header_ptr_ = header_ptr;
   data_id_ptr_ = header_pod_ptr_.MutTensorPtr<char>(FieldKey::kDataId, nullptr);
   col_num_ptr_ = header_pod_ptr_.MutTensorPtr<int32_t>(FieldKey::kColNum, nullptr);
-  dim0_valid_num_ptr_ = header_pod_ptr_.MutTensorPtr<int32_t>(FieldKey::kDim0ValidNum, nullptr);
-  dim1_valid_num_ptr_ = header_pod_ptr_.MutTensorPtr<int32_t>(FieldKey::kDim1ValidNum, nullptr);
-  dim2_valid_num_ptr_ = header_pod_ptr_.MutTensorPtr<int32_t>(FieldKey::kDim2ValidNum, nullptr);
+  dim0_valid_num_ptr_ = header_pod_ptr_.MutTensorPtr<int64_t>(FieldKey::kDim0ValidNum, nullptr);
+  dim1_valid_num_ptr_ = header_pod_ptr_.MutTensorPtr<int64_t>(FieldKey::kDim1ValidNum, nullptr);
+  dim2_valid_num_ptr_ = header_pod_ptr_.MutTensorPtr<int64_t>(FieldKey::kDim2ValidNum, nullptr);
   dptr_ = body_ptr;
   dynamic_shape_ = blob_desc->shape();
 }
@@ -48,10 +48,10 @@ void Blob::set_col_num(int32_t no, int32_t val) {
   *(col_num_ptr_ + no) = val;
 }
 
-int32_t Blob::dim0_valid_num(int32_t no) const {
+int64_t Blob::dim0_valid_num(int64_t no) const {
   CHECK_GE(no, 0);
   CHECK_LT(no, dim0_inner_shape().At(0));
-  int32_t val;
+  int64_t val;
   if (dim0_valid_num_ptr_) {
     val = dim0_valid_num_ptr_[no];
     CHECK_GE(val, 0);
@@ -61,7 +61,7 @@ int32_t Blob::dim0_valid_num(int32_t no) const {
   }
   return val;
 }
-void Blob::set_dim0_valid_num(int32_t no, int32_t val) {
+void Blob::set_dim0_valid_num(int64_t no, int64_t val) {
   CHECK_NOTNULL(dim0_valid_num_ptr_);
   CHECK_GE(no, 0);
   CHECK_LT(no, dim0_inner_shape().At(0));
@@ -70,10 +70,10 @@ void Blob::set_dim0_valid_num(int32_t no, int32_t val) {
   dim0_valid_num_ptr_[no] = val;
 }
 
-int32_t Blob::dim1_valid_num(int32_t no) const {
+int64_t Blob::dim1_valid_num(int64_t no) const {
   CHECK_GE(no, 0);
   CHECK_LT(no, blob_desc_->shape().At(0));
-  int32_t val;
+  int64_t val;
   if (dim1_valid_num_ptr_) {
     val = dim1_valid_num_ptr_[no];
     CHECK_GE(val, 0);
@@ -83,7 +83,7 @@ int32_t Blob::dim1_valid_num(int32_t no) const {
   }
   return val;
 }
-void Blob::set_dim1_valid_num(int32_t no, int32_t val) {
+void Blob::set_dim1_valid_num(int64_t no, int64_t val) {
   CHECK_NOTNULL(dim1_valid_num_ptr_);
   CHECK_GE(no, 0);
   CHECK_LT(no, blob_desc_->shape().At(0));
@@ -92,12 +92,12 @@ void Blob::set_dim1_valid_num(int32_t no, int32_t val) {
   dim1_valid_num_ptr_[no] = val;
 }
 
-int32_t Blob::dim2_valid_num(int32_t dim0_idx, int32_t dim1_idx) const {
+int64_t Blob::dim2_valid_num(int64_t dim0_idx, int64_t dim1_idx) const {
   CHECK_GE(dim0_idx, 0);
   CHECK_LT(dim0_idx, blob_desc_->shape().At(0));
   CHECK_GE(dim1_idx, 0);
   CHECK_LT(dim1_idx, blob_desc_->shape().At(1));
-  int32_t val;
+  int64_t val;
   if (dim2_valid_num_ptr_) {
     val = *(dim2_valid_num_ptr_ + dim0_idx * blob_desc_->shape().At(1) + dim1_idx);
     CHECK_GE(val, 0);
@@ -107,7 +107,7 @@ int32_t Blob::dim2_valid_num(int32_t dim0_idx, int32_t dim1_idx) const {
   }
   return val;
 }
-void Blob::set_dim2_valid_num(int32_t dim0_idx, int32_t dim1_idx, int32_t val) {
+void Blob::set_dim2_valid_num(int64_t dim0_idx, int64_t dim1_idx, int64_t val) {
   CHECK_NOTNULL(dim2_valid_num_ptr_);
   CHECK_GE(dim0_idx, 0);
   CHECK_LT(dim0_idx, blob_desc_->shape().At(0));
